+++ NEWS.md
@@ -2,6 +2,11 @@
 ShinyImgVoteR 0.1.3 (Development)
 ================
 
+## Features
+
+* Added the option to configure which voting options have a comment box via `comment_box` field in the `radio_options` configuration in config.yaml
+  - Set `comment_box: true` for a voting option to show a comment box when that option is selected
+
 ## Bug Fixes
 
 * Fixed state persistence bug where observations and comments were wronly recorded when switching vote types
@@ -13,7 +18,7 @@ ShinyImgVoteR 0.1.3 (Development)
   -  Now, sub-options are cleared when the vote type changes. Similarly, the comment field value is not recorded when the vote type has no comment field (for example, when the vote type is “Yes”).
 
 * Fixed unconfigured voting options being skipped at 3 votes
-  - Previously, voting options without explicit configuration in `voting_options_max_matching_votes` 
+  - Previously, voting options without explicit `max_matching_votes` inside `radio_options`
   were skipped after reaching 3 matching votes because there was a hardcoded default limit of 3 votes for each option
   which needed to be explicitly overridden in the configuration file.
 
@@ -59,7 +64,7 @@ ShinyImgVoteR 0.1.1
     the voting/navigation hotkeys are disabled
 
 * Added configurable maximum matching votes per screenshot across all users 
-  - It can be set via `voting_options_max_matching_votes` in config.yaml
+  - It can be set via `radio_options.<option>.max_matching_votes` in config.yaml
   - Default is 3 votes for each option if not specified
   - If maximum matching votes reached for a screenshot, it will be skipped for future users
   - Skipped screenshots are logged in user annotations file with reason "skipped - max matching votes (x) for option (y) reached" in the agreement column